refactor: extract shared field management methods to trait
