+++ src/Livewire/ManageFieldsWithoutSections.php
@@ -12,13 +12,11 @@ use Filament\Forms\Contracts\HasForms;
 use Filament\Support\Enums\Size;
 use Filament\Support\Enums\Width;
 use Illuminate\Contracts\View\View;
-use Illuminate\Database\Eloquent\Collection;
-use Livewire\Attributes\Computed;
-use Livewire\Attributes\On;
 use Livewire\Component;
 use Relaticle\CustomFields\CustomFields;
 use Relaticle\CustomFields\Filament\Management\Schemas\FieldForm;
 use Relaticle\CustomFields\Livewire\Concerns\CreatesCustomFields;
+use Relaticle\CustomFields\Livewire\Concerns\ManagesFields;
 use Relaticle\CustomFields\Models\CustomFieldSection;
 
 /**
@@ -29,37 +27,11 @@ final class ManageFieldsWithoutSections extends Component implements HasActions,
     use CreatesCustomFields;
     use InteractsWithActions;
     use InteractsWithForms;
+    use ManagesFields;
 
     public string $entityType;
 
     public CustomFieldSection $section;
-
-    #[Computed]
-    public function fields(): Collection
-    {
-        return $this->section->fields()->withDeactivated()->orderBy('sort_order')->get();
-    }
-
-    #[On('field-width-updated')]
-    public function fieldWidthUpdated(int|string $fieldId, int $width): void
-    {
-        $model = CustomFields::newCustomFieldModel();
-        $model->where($model->getKeyName(), $fieldId)->update(['width' => $width]);
-
-        $this->section->refresh();
-    }
-
-    #[On('field-deleted')]
-    public function fieldDeleted(): void
-    {
-        $this->section->refresh();
-    }
-
-    #[On('fields-reordered')]
-    public function fieldsReordered(): void
-    {
-        unset($this->fields);
-    }
 
     public function updateFieldsOrder(array $fields): void
     {
